MfgTool.exe - Release 1.6.1.000 : ENGR00117829: MFGTool.exe will crash when ucl.xml contains a commented out field in the <LIST/> node.
DoListThreadProc() was trying to process all <LIST/> nodes instead of all <LIST/> <CMD/> nodes. When some commands were commented out, the count was off and the app was trying to get a <CMD/> node that was not in the list. I.e. the accessed index was bigger than the array.
	  
Changed the loop to only process <CMD/> nodes instead of all child nodes.

git-svn-id: http://psg-host.svn-tx30.am.freescale.net/svn/psg-host/branches/dev/0004@304 41b5e5e7-094d-0410-8aea-ef766df42c58