Allow different config per device when programming multiple devices
Before this patch, when programming multiple devices with different
configuration (mac addresses fx.), we would hit an issue where all
devices would get the same configuration.
The reason for this was that even though the configuration was updated
for each device, the output of the update was stored in a single class
variable (m_bodyString), which was getting overwritten by the lastest
update.

So this patch:
1. Sets up a template variable (m_templateBodyString) including
getters/setters, which is used to store the original XML.
2. Changes the m_bodyString to be a map, based on the WndIndex of the
device and update references to it.
3. Make sure that nothing is directly using the class variable
(m_bodyString), but instead all operations use the getters and setters.

Signed-off-by: Paolo La Camera <paolo@airtame.com>