Fix polling loop
The UtpSenseCode check wasn't done in the correct byte order so
SendCommand never actually went into the polling loop.

After fixing that it turned out it would only ever do a loop once as
bCSWStatus was never 1 in response to a poll request. To fix that simply
always go around the loop again after a poll to check the current Sense
response.

Signed-off-by: Sjoerd Simons <sjoerd.simons@collabora.co.uk>