build: Only create gitversion.h on success
So far gitversion.h would always be created in case of errors. Sometimes
it would be empty, sometimes it would just contain "lib" without a
version number. The later leads to segfaults at run time.

This makes sure we only move the gitversion.h into place if we could the
script succeeded.

Closes: #98