Allow reading the SUDO_USER's uuu script whenthe snap is run as root
Due to snap's security confinement, by default only scripts under the
home dir is readable, which is /root when the snap is run via sudo
and is not desirable.

This patch introduces the read attribute to the home interface plug
which allows access to files under the user's home directory
(as long as traditional file permissions allow).

Signed-off-by: 林博仁(Buo-ren Lin) <Buo.Ren.Lin@gmail.com>