Add 'upload' fastboot command support
Some bootloader implementations support the 'upload' command, which allows for data
to be transmitted from UBOOT -> UUU.  This command is necessary when performing advanced
flashing operations, such as configuring the security features on i.MX SoC's.

Example of where this command would be utilized: uploading the 'mppubk' from
UBOOT so attestation keys and certificates can be encrypted before they are pushed
and burned to secure storage in a 'closed chip' configuration.