Specify 'ios::binary' in ofstream initialization
When executed on a Windows platform, any data that contains a line feed character
('\n') will automatically have a carriage return ('\r') added before the line feed, unless
the ofstream is opened with the 'ios::binary' flag.  Adding carriage returns to important
data retrieved from the fastboot buffer (e.g. MPPUBK) is not ideal, as it can interfere with
how that data is used.