Skip sleep in usb_add
The 200ms sleep in usb_add is run once per attached usb device. On a
system with many usb devices, this can result in a wait time of several
seconds, even when the device waited for is already connected.

Remove the sleep, since I don't see any adverse effects of doing so.

Signed-off-by: Torleiv Sundre <torleiv@huddly.com>