Fix issues when m_skip is not zero in fastboot
m_skip is expected to skip the source file, but when it's not zero,
it will be mistakenly used to seek the target device as well.

When both m_skip and m_seek are not zero, the offset of target device
will be m_skip + m_seek, but should be m_seek.

So separate them when accessing source file and target device in fastboot.

Also Change crcblock parameter of FBCRC command to each and add FBWrite
command to help info.