parser: fix behavior with undefined environment variable on linux
The @ operator is used to expand an environment variable in the script.

When the variable does not exists in the environment, the behavior
between linux and windows differs. On linux this is failing with an
error message "variable '...' is not defined", on windows the variable
is evaluated to an empty string.

This commit unifies the behavior and evaluates the undefined variable
to an empty string without failing.