libuuu: hidreport: fix write size
Currently write() always sends buffers of 'm_size_out' size in case of
report_id != 1. This may cause an issue for the last buffer if not
aligned to 'm_size_out'.

Fix this by send only the required size of bytes by using 's'. This also
simplifies the call a lot.

Signed-off-by: Marco Felsch <m.felsch@pengutronix.de>