libuuu: sdp: SDPBootCmd: add barebox autodetection
We do already support the -barebox switch, so the user can force the
barebox image handling. Add a autodetection mechanism to make it even
more user friendly. The autodetection is based on the fact, that barebox
does add an "barebox" string at offset 0x20 to the image. If this string
is found we can assume that user provided file is an barebox image.

Signed-off-by: Marco Felsch <m.felsch@pengutronix.de>