libuuu/usbhotplug.cpp: don't limit retry logic to windows
When using the udev rules to avoid use of sudo, we have sometimes
observed uuu spuriously failing because it sees the device and
attempts to open it, before the udev daemon has had a chance to apply
its rules and grant access to the current user (via the 'uaccess'
tag).

The error in that case is -EACCES which libusb translates to
LIBUSB_ERROR_ACCESS, so extend the retry logic to that case. Also,
since I have seen the open() succeed on the 5th and last attempt, bump
the number of retries to 10, for a total of around 2 seconds instead
of just 1.