libuuu: fastboot.cpp: check if logical-block-size is converted
In case, if "logical-block-size" is not defined and Fastboot device
sends empty "INFO" message (for example, barebox does so),
"str_to_uint32" gets empty string and returns UINT32_MAX,
which is used as block size.

This huge block size is obviously incorrect and image is flashed
incorrectly.

To prevent such a behaviour, check if a conversion succeeded.

Signed-off-by: Andrei Lalaev <andrey.lalaev@gmail.com>