Fixes #476: Accessing freed memory in fastboot.c
The original code takes a pointer (sparse_header * pfile) to a DataBuffer
that is freed in subsequent calls to the request_data() function of the
underlying FileBuffer. However, the pointer is still used to determine
various aspects of the sparse file later. These may be already changed if
the memory is reused, which leads to invalid values and trashed transfers.

Fix is to store a sparse_header copy as separate variables.

Reported-by: Henrik Hedberg <henrik.hedberg@haltian.com>
Signed-off-by: Frank Li <Frank.Li@nxp.com>